Split the coordinator out of wfrac/device.py into coordinator.py
Home Assistant core expects the DataUpdateCoordinator in coordinator.py;
the WF-RAC protocol client stays behind in wfrac/.
