+++ custom_components/mitsubishi_wf_rac/binary_sensor.py
@@ -14,7 +14,7 @@ from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
 from . import MitsubishiWfRacConfigEntry
 from .entity import WfRacEntity
-from .wfrac.device import Device
+from .coordinator import Device
 from .wfrac.error_codes import describe_error_code
 from .const import DOMAIN
