+++ custom_components/mitsubishi_wf_rac/button.py
@@ -12,7 +12,7 @@ from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
 from . import MitsubishiWfRacConfigEntry
 from .entity import WfRacEntity
-from .wfrac.device import Device
+from .coordinator import Device
 from .const import DOMAIN, SIGNAL_SET_ENERGY_TOTAL
 
 _LOGGER = logging.getLogger(__name__)
