+++ custom_components/mitsubishi_wf_rac/const.py
@@ -22,7 +22,7 @@ CONF_AIRCO_ID = "airco_id"
 # that predate v5. Nothing outside the migration reads it.
 CONF_AVAILABILITY_CHECK = "availability_check"
 # Consecutive failed polls before the device is reported unavailable; floored
-# at wfrac/device.py's AVAILABILITY_FAILURE_LIMIT_MIN.
+# at coordinator.py's AVAILABILITY_FAILURE_LIMIT_MIN.
 CONF_AVAILABILITY_RETRY_LIMIT = "availability_retry_limit"
 # Gates all outbound internet traffic (as opposed to local-network device
 # polling) - the manufacturer's getFirmware endpoint. Off by default: unlike
