+++ tests/integration/test_coordinator.py
@@ -17,8 +17,8 @@ from homeassistant.helpers import issue_registry as ir
 from pytest_homeassistant_custom_component.common import MockConfigEntry
 
 from custom_components.mitsubishi_wf_rac.const import DOMAIN
-from custom_components.mitsubishi_wf_rac.wfrac import device as device_module
-from custom_components.mitsubishi_wf_rac.wfrac.device import (
+from custom_components.mitsubishi_wf_rac import coordinator as coordinator_module
+from custom_components.mitsubishi_wf_rac.coordinator import (
     AVAILABILITY_FAILURE_LIMIT_MIN,
     Device,
 )
@@ -90,13 +90,13 @@ def _shorten_service_data_timing(monkeypatch, offset_ms: int = 5) -> None:
     test can wait out.
     """
     monkeypatch.setattr(
-        device_module, "UPDATE_CONSOLIDATION_PERIOD", timedelta(milliseconds=5)
+        coordinator_module, "UPDATE_CONSOLIDATION_PERIOD", timedelta(milliseconds=5)
     )
     monkeypatch.setattr(
-        device_module, "SERVICE_DATA_REQUEST_OFFSET", timedelta(milliseconds=offset_ms)
+        coordinator_module, "SERVICE_DATA_REQUEST_OFFSET", timedelta(milliseconds=offset_ms)
     )
     monkeypatch.setattr(
-        device_module, "SERVICE_DATA_RETRY_DELAY", timedelta(milliseconds=5)
+        coordinator_module, "SERVICE_DATA_RETRY_DELAY", timedelta(milliseconds=5)
     )
 
 
@@ -150,7 +150,7 @@ async def test_update_transient_unreachable_is_debug_only(device, caplog):
     bare connection failure there is nothing to re-register against, and the
     hourly WiFi restart these modules do would make it a recurring no-op.
     """
-    caplog.set_level("DEBUG", logger=device_module.__name__)
+    caplog.set_level("DEBUG", logger=coordinator_module.__name__)
     device._api.get_aircon_stats.return_value = _stats_response(ON_COOL_PAYLOAD)
     await device.update()
     caplog.clear()
@@ -161,7 +161,7 @@ async def test_update_transient_unreachable_is_debug_only(device, caplog):
 
     assert device.available is True
     device._api.update_account_info.assert_not_awaited()
-    records = [r for r in caplog.records if r.name == device_module.__name__]
+    records = [r for r in caplog.records if r.name == coordinator_module.__name__]
     assert not [r for r in records if r.levelname == "WARNING"]
     assert len(records) == 1
     assert records[0].levelname == "DEBUG"
@@ -174,7 +174,7 @@ async def test_update_transient_unreachable_is_debug_only(device, caplog):
 
 
 async def test_update_sustained_unreachable_logs_one_transition(device, caplog):
-    caplog.set_level("DEBUG", logger=device_module.__name__)
+    caplog.set_level("DEBUG", logger=coordinator_module.__name__)
     device._api.get_aircon_stats.return_value = _stats_response(ON_COOL_PAYLOAD)
     await device.update()
     caplog.clear()
@@ -190,7 +190,7 @@ async def test_update_sustained_unreachable_logs_one_transition(device, caplog):
     warnings = [
         r
         for r in caplog.records
-        if r.name == device_module.__name__ and r.levelname == "WARNING"
+        if r.name == coordinator_module.__name__ and r.levelname == "WARNING"
     ]
     assert len(warnings) == 1
     assert "is unavailable after 3 failed polls" in warnings[0].message
@@ -198,7 +198,7 @@ async def test_update_sustained_unreachable_logs_one_transition(device, caplog):
     assert sum(
         r.exc_info is not None
         for r in caplog.records
-        if r.name == device_module.__name__ and r.levelname == "DEBUG"
+        if r.name == coordinator_module.__name__ and r.levelname == "DEBUG"
     ) == 1
 
 
@@ -214,14 +214,14 @@ async def test_update_initially_unreachable_logs_threshold_once(device, caplog):
     warnings = [
         r
         for r in caplog.records
-        if r.name == device_module.__name__ and r.levelname == "WARNING"
+        if r.name == coordinator_module.__name__ and r.levelname == "WARNING"
     ]
     assert len(warnings) == 1
     assert "is unavailable after 3 failed polls" in warnings[0].message
 
 
 async def test_update_recovery_is_logged_once(device, caplog):
-    caplog.set_level("INFO", logger=device_module.__name__)
+    caplog.set_level("INFO", logger=coordinator_module.__name__)
     device._api.get_aircon_stats.side_effect = AirconConnectionError("no route")
     for _ in range(3):
         await device.update()
@@ -235,7 +235,7 @@ async def test_update_recovery_is_logged_once(device, caplog):
     recoveries = [
         r
         for r in caplog.records
-        if r.name == device_module.__name__
+        if r.name == coordinator_module.__name__
         and r.levelname == "INFO"
         and "is available again" in r.message
     ]
@@ -418,7 +418,7 @@ async def test_set_airco_lock_prevents_stale_snapshot_race(device):
 
 
 async def test_async_queue_command_coalesces_into_one_send(device, monkeypatch):
-    monkeypatch.setattr(device_module, "UPDATE_CONSOLIDATION_PERIOD", timedelta(milliseconds=5))
+    monkeypatch.setattr(coordinator_module, "UPDATE_CONSOLIDATION_PERIOD", timedelta(milliseconds=5))
     device._api.get_aircon_stats.return_value = _stats_response(OFF_PAYLOAD)
     await device.update()
     device._api.send_airco_command = AsyncMock(side_effect=_echo_send_airco_command)
@@ -434,7 +434,7 @@ async def test_async_queue_command_coalesces_into_one_send(device, monkeypatch):
 
 
 async def test_async_queue_command_notifies_listeners(device, monkeypatch):
-    monkeypatch.setattr(device_module, "UPDATE_CONSOLIDATION_PERIOD", timedelta(milliseconds=5))
+    monkeypatch.setattr(coordinator_module, "UPDATE_CONSOLIDATION_PERIOD", timedelta(milliseconds=5))
     device._api.get_aircon_stats.return_value = _stats_response(OFF_PAYLOAD)
     await device.update()
     device._api.send_airco_command = AsyncMock(side_effect=_echo_send_airco_command)
@@ -456,7 +456,7 @@ async def test_async_queue_command_notifies_listeners(device, monkeypatch):
 
 
 async def test_async_queue_command_notifies_listeners_even_on_failure(device, monkeypatch):
-    monkeypatch.setattr(device_module, "UPDATE_CONSOLIDATION_PERIOD", timedelta(milliseconds=5))
+    monkeypatch.setattr(coordinator_module, "UPDATE_CONSOLIDATION_PERIOD", timedelta(milliseconds=5))
     device._api.get_aircon_stats.return_value = _stats_response(OFF_PAYLOAD)
     await device.update()
     device._api.send_airco_command = AsyncMock(
@@ -484,7 +484,7 @@ async def test_home_leave_mode_status_request_does_not_swallow_a_queued_command(
     change) went out unset and was silently ignored by the unit. Sending the
     status request directly through set_airco() keeps it out of that merge.
     """
-    monkeypatch.setattr(device_module, "UPDATE_CONSOLIDATION_PERIOD", timedelta(milliseconds=5))
+    monkeypatch.setattr(coordinator_module, "UPDATE_CONSOLIDATION_PERIOD", timedelta(milliseconds=5))
     device._api.get_aircon_stats.return_value = _stats_response(OFF_PAYLOAD)
     await device.update()
     sent = []
@@ -632,7 +632,7 @@ async def test_update_does_not_check_firmware_when_disabled_by_default(device, m
     # firmware_update_check option (see const.py's CONF_FIRMWARE_UPDATE_CHECK).
     assert device.firmware_update_check_enabled is False
     fetch = AsyncMock(return_value={"wireless": "026", "mcu": "200"})
-    monkeypatch.setattr(device_module, "fetch_latest_firmware", fetch)
+    monkeypatch.setattr(coordinator_module, "fetch_latest_firmware", fetch)
     device._api.get_aircon_stats.return_value = _stats_response_with_firmware(
         ON_COOL_PAYLOAD, "WF-RAC-HTTPS", "025"
     )
@@ -648,7 +648,7 @@ async def test_update_does_not_check_firmware_when_disabled_by_default(device, m
 async def test_update_detects_available_firmware_update(device, monkeypatch):
     device._firmware_update_check_enabled = True
     fetch = AsyncMock(return_value={"wireless": "026", "mcu": "200"})
-    monkeypatch.setattr(device_module, "fetch_latest_firmware", fetch)
+    monkeypatch.setattr(coordinator_module, "fetch_latest_firmware", fetch)
     device._api.get_aircon_stats.return_value = _stats_response_with_firmware(
         ON_COOL_PAYLOAD, "WF-RAC-HTTPS", "025"
     )
@@ -668,7 +668,7 @@ async def test_update_does_not_flag_downgrade_or_equal_version_as_update(device,
     # reported as an available update.
     device._firmware_update_check_enabled = True
     fetch = AsyncMock(return_value={"wireless": "025", "mcu": "200"})
-    monkeypatch.setattr(device_module, "fetch_latest_firmware", fetch)
+    monkeypatch.setattr(coordinator_module, "fetch_latest_firmware", fetch)
     device._api.get_aircon_stats.return_value = _stats_response_with_firmware(
         ON_COOL_PAYLOAD, "WF-RAC-HTTPS", "025"
     )
@@ -682,7 +682,7 @@ async def test_update_does_not_flag_downgrade_or_equal_version_as_update(device,
 async def test_update_firmware_check_is_rate_limited(device, monkeypatch):
     device._firmware_update_check_enabled = True
     fetch = AsyncMock(return_value={"wireless": "026", "mcu": "200"})
-    monkeypatch.setattr(device_module, "fetch_latest_firmware", fetch)
+    monkeypatch.setattr(coordinator_module, "fetch_latest_firmware", fetch)
     device._api.get_aircon_stats.return_value = _stats_response_with_firmware(
         ON_COOL_PAYLOAD, "WF-RAC-HTTPS", "025"
     )
@@ -698,7 +698,7 @@ async def test_update_firmware_check_is_rate_limited(device, monkeypatch):
 async def test_update_firmware_check_failure_leaves_state_unknown(device, monkeypatch):
     device._firmware_update_check_enabled = True
     fetch = AsyncMock(return_value=None)
-    monkeypatch.setattr(device_module, "fetch_latest_firmware", fetch)
+    monkeypatch.setattr(coordinator_module, "fetch_latest_firmware", fetch)
     device._api.get_aircon_stats.return_value = _stats_response_with_firmware(
         ON_COOL_PAYLOAD, "WF-RAC-HTTPS", "025"
     )
@@ -714,7 +714,7 @@ async def test_update_firmware_check_failure_leaves_state_unknown(device, monkey
 
 
 async def test_update_does_not_request_service_data_without_active_entities(device, monkeypatch):
-    monkeypatch.setattr(device_module, "UPDATE_CONSOLIDATION_PERIOD", timedelta(milliseconds=5))
+    monkeypatch.setattr(coordinator_module, "UPDATE_CONSOLIDATION_PERIOD", timedelta(milliseconds=5))
     device._api.get_aircon_stats.return_value = _stats_response(ON_COOL_PAYLOAD)
     device._api.send_airco_command = AsyncMock(side_effect=_echo_send_airco_command)
 
@@ -812,7 +812,7 @@ async def test_add_account_returns_none_on_api_error(device):
 
 def _issue(device):
     return ir.async_get(device._hass).async_get_issue(
-        DOMAIN, device_module.registration_full_issue_id(device.config_entry.entry_id)
+        DOMAIN, coordinator_module.registration_full_issue_id(device.config_entry.entry_id)
     )
 
 
@@ -1013,7 +1013,7 @@ async def test_service_data_survives_a_poll_that_answered_early(device, monkeypa
     device._api.get_aircon_stats.return_value = _stats_response(ON_COOL_PAYLOAD)
     device._api.send_airco_command = AsyncMock(side_effect=_echo_send_airco_command)
     device._last_service_data_request = datetime.now() - (
-        device_module.SERVICE_DATA_REQUEST_INTERVAL - timedelta(milliseconds=100)
+        coordinator_module.SERVICE_DATA_REQUEST_INTERVAL - timedelta(milliseconds=100)
     )
 
     await device.update()
@@ -1036,7 +1036,7 @@ async def test_async_update_data_wraps_exception_in_update_failed(device):
 async def test_coordinator_tracks_transient_failure_without_regular_log_noise(
     device, caplog
 ):
-    caplog.set_level("DEBUG", logger=device_module.__name__)
+    caplog.set_level("DEBUG", logger=coordinator_module.__name__)
     device._api.get_aircon_stats.return_value = _stats_response(ON_COOL_PAYLOAD)
     await device.async_refresh()
     caplog.clear()
@@ -1087,7 +1087,7 @@ async def test_coordinator_notifies_when_device_reaches_unavailable_threshold(de
 
 
 async def test_coordinator_still_logs_unexpected_failures(device, caplog):
-    caplog.set_level("DEBUG", logger=device_module.__name__)
+    caplog.set_level("DEBUG", logger=coordinator_module.__name__)
 
     async def _boom():
         raise RuntimeError("unexpected")
@@ -1102,8 +1102,8 @@ async def test_coordinator_still_logs_unexpected_failures(device, caplog):
 async def test_async_update_data_counts_timeouts_as_connection_failures(
     device, monkeypatch, caplog
 ):
-    monkeypatch.setattr(device_module, "POLL_TIMEOUT", timedelta(milliseconds=10))
-    caplog.set_level("DEBUG", logger=device_module.__name__)
+    monkeypatch.setattr(coordinator_module, "POLL_TIMEOUT", timedelta(milliseconds=10))
+    caplog.set_level("DEBUG", logger=coordinator_module.__name__)
     device._api.get_aircon_stats.return_value = _stats_response(ON_COOL_PAYLOAD)
     await device.async_refresh()
     original_update = device.update
@@ -1124,7 +1124,7 @@ async def test_async_update_data_counts_timeouts_as_connection_failures(
     warnings = [
         record
         for record in caplog.records
-        if record.name == device_module.__name__ and record.levelname == "WARNING"
+        if record.name == coordinator_module.__name__ and record.levelname == "WARNING"
     ]
     assert len(warnings) == 1
     assert "is unavailable after 3 failed polls" in warnings[0].message
@@ -1221,7 +1221,7 @@ async def test_service_data_expires_when_nothing_fresh_arrives(device):
     """
     device._airco.CompressorFrequency = 40.0
     device._last_service_data_response = datetime.now() - (
-        device_module.SERVICE_DATA_MAX_AGE + timedelta(seconds=1)
+        coordinator_module.SERVICE_DATA_MAX_AGE + timedelta(seconds=1)
     )
     new_airco = Aircon()
 
@@ -1234,7 +1234,7 @@ async def test_fresh_service_data_restarts_the_clock(device):
     device._airco.CompressorFrequency = 40.0
     device._airco.HotGasTemp = 50.0
     device._last_service_data_response = datetime.now() - (
-        device_module.SERVICE_DATA_MAX_AGE + timedelta(seconds=1)
+        coordinator_module.SERVICE_DATA_MAX_AGE + timedelta(seconds=1)
     )
     new_airco = Aircon()
     new_airco.CompressorFrequency = 45.0  # this poll carried the segments
@@ -1250,10 +1250,10 @@ async def test_expiring_service_data_warns_once_and_reports_the_recovery(
     device, caplog
 ):
     """The refusals themselves are routine; running out of values is not."""
-    caplog.set_level("DEBUG", logger=device_module.__name__)
+    caplog.set_level("DEBUG", logger=coordinator_module.__name__)
     device._airco.CompressorFrequency = 40.0
     device._last_service_data_response = datetime.now() - (
-        device_module.SERVICE_DATA_MAX_AGE + timedelta(seconds=1)
+        coordinator_module.SERVICE_DATA_MAX_AGE + timedelta(seconds=1)
     )
 
     for _ in range(3):
@@ -1278,7 +1278,7 @@ async def test_service_data_that_never_arrived_stays_quiet_until_it_is_due(
     device, caplog
 ):
     """A unit asked for the first time has nothing to lose yet."""
-    caplog.set_level("DEBUG", logger=device_module.__name__)
+    caplog.set_level("DEBUG", logger=coordinator_module.__name__)
     device._last_service_data_request = datetime.now()
 
     device._carry_forward_service_data(Aircon())
