+++ tests/integration/test_entity_platforms.py
@@ -29,7 +29,7 @@ from custom_components.mitsubishi_wf_rac.const import (
     SWING_HORIZONTAL_MODE_TRANSLATION,
     SWING_MODE_TRANSLATION,
 )
-from custom_components.mitsubishi_wf_rac.wfrac.device import Device
+from custom_components.mitsubishi_wf_rac.coordinator import Device
 from custom_components.mitsubishi_wf_rac.wfrac.models.aircon import AirconCommands, HomeLeaveModeSetting
 
 from ..unit.live_captures import LIVE_CAPTURES
